+++ protocols/index.js
@@ -41,11 +41,13 @@ import rfactor from './rfactor.js'
 import samp from './samp.js'
 import satisfactory from './satisfactory.js'
 import savage2 from './savage2.js'
+import soldat from './soldat.js'
 import starmade from './starmade.js'
 import starsiege from './starsiege.js'
 import teamspeak2 from './teamspeak2.js'
 import teamspeak3 from './teamspeak3.js'
 import terraria from './terraria.js'
+import toxikk from './toxikk.js'
 import tribes1 from './tribes1.js'
 import tribes1master from './tribes1master.js'
 import unreal2 from './unreal2.js'
@@ -68,7 +70,6 @@ export {
   armagetron, ase, asa, assettocorsa, battlefield, buildandshoot, cs2d, discord, doom3, eco, epic, factorio, farmingsimulator, ffow,
   fivem, gamespy1, gamespy2, gamespy3, geneshift, goldsrc, gtasao, hexen2, jc2mp, kspdmp, mafia2mp, mafia2online, minecraft,
   minecraftbedrock, minecraftvanilla, minetest, mumble, mumbleping, nadeo, openttd, palworld, quake1, quake2, quake3, rfactor, ragemp, samp,
-  satisfactory, savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, tribes1, tribes1master, unreal2, ut3, valve,
-  vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp, dayz, theisleevrima, xonotic, altvmp, vintagestorymaster,
-  vintagestory
+  satisfactory, soldat, savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, toxikk, tribes1, tribes1master, unreal2, ut3, valve,
+  vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp, dayz, theisleevrima, xonotic, altvmp, vintagestorymaster, vintagestory
 }